Moving configuration file.
